requirements.txt
django == 3.2.12
channels==3.0.4
channels-redis==3.3.1
rpi_ws281x
adafruit-circuitpython-neopixel
adafruit-blinka